Chore #4 : Package refactoring
